+++ utils/fetch-schema/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @flatfile/util-fetch-schema
 
+## 0.1.0
+
+### Minor Changes
+
+- 30981b2: Dependency updates
+
 ## 0.0.3
 
 ### Patch Changes